add weather for 5 days and description on now weather

+++ lesson_4_part_2/lib/domains/weather_model/weather_data_model.dart
@@ -120,7 +120,7 @@ class Main {
 
   Main.fromJson(Map<String, dynamic> json) {
     temp = json['temp'].toDouble();
-    feelsLike = json['feels_like'];
+    feelsLike = json['feels_like'].toDouble();
     tempMin = json['temp_min'].toDouble();
     tempMax = json['temp_max'].toDouble();
   }
